Conditional testing based on PHP-version.
