+++ .travis.yml
@@ -25,5 +25,3 @@ matrix:
       env: ILLUMINATE_VERSION=5.8.*
     - php: 7.3
       env: ILLUMINATE_VERSION=5.8.*
-    - php: nightly
-      env: ILLUMINATE_VERSION=5.8.*
