+++ phpunit.xml
@@ -11,8 +11,8 @@
 >
     <testsuites>
         <testsuite name="Package Test Suite">
-            <directory suffix=".php">./tests/</directory>
-            <exclude>./tests/app/</exclude>
+            <file phpVersion="7.1" phpVersionOperator="lt">./tests/PHP70.php</file>
+            <file phpVersion="7.1" phpVersionOperator="ge">./tests/PHP71.php</file>
         </testsuite>
     </testsuites>
 </phpunit>
